% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmd.R
\name{asmd}
\alias{asmd}
\title{asmd}
\usage{
asmd(numbers)
}
\arguments{
\item{numbers}{Input vector of numbers to add together, subtract from 0, multiply together or divide 1 by}
}
\value{
Returns 4 componants, add, divide, multiple and subtract that contain the return values
}
\description{
asmd
}
\examples{
asmd(c(1,2,3,4,5))
}
\keyword{mathematics}
