% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assignment.R
\docType{package}
\name{assignment}
\alias{assignment}
\title{Kaya Identity Equation Package}
\description{
contains one function kaya Identity which expresses yearly CO2 emissions
}
\author{
Luyang Chu \email{luyang.chu@gmail.com}
}
