% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{CSV File Reader.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{This should be a character value. If the file is in the present working directory, then the
name of the file which is to be read, otherwise the full path to acces the file (with file name).}
}
\value{
The data present in the file.This is an object of class 'tbl_df'.
}
\description{
Reads a csv file, given the filename.
}
\note{
Requires 'readr' and 'dplyr' packages to be loaded first.
}
\examples{
## Considering your data is in the present working directory
mydata <- fars_read("accident_2013.csv.bz2")
str(mydata)

}
