% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize 'Year' & 'Month' variables from the data.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The year for which the Year & Month variables are to be Summarized from the data. This should
be integer.}
}
\value{
The count of number of observations present in the file (in case year = 2013, 2014, 2015) for each month.
        This is an object of class 'tbl_df'. In case the file is not present corresponding to the year supplied
         in the argument, it throws an error.
}
\description{
Summarizes Year & Month variables from the data, if the file is present in the working directory
            (in case year = 2013, 2014, 2015), otherwise prints 'Invalid year'
}
\note{
Requires 'readr', 'dplyr' and 'tidyr' packages to be loaded first.
}
\examples{
## Considering your data is in the present working directory
## The following would work fine and would return the subsetted data.
mydata <- fars_summarize_years(2014)
str(mydata)

}
