% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_first_100_primes.R
\name{in_first_100_primes}
\alias{in_first_100_primes}
\title{first 100 primes}
\usage{
in_first_100_primes(number)
}
\arguments{
\item{number}{Number to check if its in the first 100 primes}
}
\value{
Boolean response True if the number provided is in the first 100 prime numbers
}
\description{
first 100 primes
}
\examples{
in_first_100_primes(number = 15)
}
