% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya_identity_equation.R
\name{kaya_identity_equation}
\alias{kaya_identity_equation}
\title{create Kaya equation as a function}
\usage{
kaya_identity_equation(pop, gdp, enInt, carbInt, output_type = "CO2")
}
\arguments{
\item{pop}{Numeric}

\item{gdp}{Numeric}

\item{enInt}{Numeric}

\item{output_type}{String}

\item{carnInt}{Numeric}
}
\value{
Numeric
}
\description{
more detailed description.
}
\examples{
kaya_equation.germany <- kaya_equation(82.4, 44, 5, 0.05)

}
