% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{File Name Generator.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year for which the filename has to be generated.This should be an integer value.}
}
\value{
The filename in the above mentioned format for the year which has been supplied. (An object of class
        character).
}
\description{
Generates a filename (character) in the format similar to other files given as a part of the 1st
              assignment of the course 'Building R Packages" on Coursera.
}
\examples{
myfilename <- make_filename(2020)
str(myfilename)
}
