% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_library.R
\name{mean_variety}
\alias{mean_variety}
\title{mean_variety}
\usage{
mean_variety(numbers)
}
\arguments{
\item{numbers}{Input vector of numbers to determine the different means of.}
}
\value{
Returns 3 componants, the harmonic mean (harmonic_mean), the geometric mean (geometric_mean) and the standard mean (standard_mean)
}
\description{
mean_variety
}
\examples{
mean_variety(numbers = c(1,2,3,4,5))
}
\keyword{mean}
