% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\name{new_cycling_classification}
\alias{new_cycling_classification}
\title{constructor of a cycling_classification instance}
\usage{
new_cycling_classification(network, classified_lines, summary_stats)
}
\arguments{
\item{network}{A \code{cycling_network} instance.}

\item{classified_lines}{An sf object with added \code{infra_type}.}

\item{summary_stats}{A data.frame with infrastructure summary statistics.}
}
\value{
A \code{cycling_classification} object.
}
\description{
Extends a cycling_network with classification and summary statistics.
}
