% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\name{new_cycling_network}
\alias{new_cycling_network}
\title{Constructor for cycling_network instance}
\usage{
new_cycling_network(city, sf_lines)
}
\arguments{
\item{city}{Character string. Name of the city.}

\item{sf_lines}{An sf object with LINESTRING geometries.}
}
\value{
A \code{cycling_network} object.
}
\description{
Creates a \code{cycling_network} instance containing spatial cycling infrastructure
extracted from OpenStreetMap.
}
