% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\name{plot.cycling_classification}
\alias{plot.cycling_classification}
\title{Plot cycling infrastructure classification map}
\usage{
\method{plot}{cycling_classification}(x, ...)
}
\arguments{
\item{x}{A \code{cycling_classification} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Visualises cycling infrastructure classified by safety level.
}
