% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\name{plot.cycling_network}
\alias{plot.cycling_network}
\title{Plot method for cycling_network instance}
\usage{
\method{plot}{cycling_network}(x, ...)
}
\arguments{
\item{x}{A \code{cycling_network} instance.}

\item{...}{Additional arguments (ignored).}
}
\description{
Plot method for cycling_network instance
}
