% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primes}
\alias{primes}
\title{Data file of first 100 primes}
\format{
A vector with 100 values of prime numbers:
}
\source{
{google copy and paste job.}
}
\usage{
primes
}
\description{
Contains the raw numbers
}
\keyword{datasets}
