% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\name{print.cycling_classification}
\alias{print.cycling_classification}
\title{Print method for cycling_classification instance}
\usage{
\method{print}{cycling_classification}(x, ...)
}
\arguments{
\item{x}{A \code{cycling_classification} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for cycling_classification instance
}
